+++ packages/wallets/keepkey/src/keepkey.ts
@@ -255,11 +255,11 @@ const connectKeepkey =
     await checkAndLaunch();
 
     if (!paths) paths = [];
-    console.log('keepkey: paths: ', paths);
+    // console.log('keepkey: paths: ', paths);
     // Only build this once for all assets
     const keepKeySdk = await KeepKeySdk.create(keepkeyConfig);
-    console.log('keepKeySdk: ', keepKeySdk);
-    console.log('keepkeyConfig.apiKey: ', keepkeyConfig.apiKey);
+    // console.log('keepKeySdk: ', keepKeySdk);
+    // console.log('keepkeyConfig.apiKey: ', keepkeyConfig.apiKey);
     let features = await keepKeySdk.system.info.getFeatures();
     // console.log('features: ', features);